+++ src/components/climate/ClimateModal.js
@@ -7,18 +7,17 @@ import {
     Button,
     ButtonStrip,
 } from '@dhis2/ui'
-import React, { useState, useEffect } from 'react'
+import React, { useState } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { closeClimatePanel } from '../../actions/climate.js'
-import { getTimeSeries } from '../../util/earthEngine.js'
 import { Tab, Tabs } from '../core/index.js'
 import AirQuality from './AirQuality.js'
 import ClimateChange from './ClimateChange.js'
 import Forecast from './Forecast.js'
 import Precipitation from './Precipitation.js'
 import styles from './styles/Modal.module.css'
-import Temperature from './Temperature.js'
-// import data from './data-monthly.js'
+import TemperatureTab from './TemperatureTab.js'
+import useClimateData from '../../hooks/useClimateData.js'
 
 const FORECAST = 'forecast'
 const PRECIPITATION = 'precipitation'
@@ -26,36 +25,14 @@ const TEMPERATURE = 'temperature'
 const CLIMATE = 'climate'
 const AIR = 'air'
 
-const config = {
-    datasetId: 'ECMWF/ERA5_LAND/MONTHLY_AGGR',
-    band: [
-        'temperature_2m',
-        'temperature_2m_min',
-        'temperature_2m_max',
-        'total_precipitation_sum',
-        'total_precipitation_min',
-        'total_precipitation_max',
-    ],
-    filters: [
-        {
-            type: 'date',
-            arguments: ['$1', '$2'],
-        },
-    ],
-}
-
 const ClimateModal = () => {
     const [tab, setTab] = useState(FORECAST)
-    const [data, setData] = useState()
     const feature = useSelector((state) => state.climate)
     const dispatch = useDispatch()
 
     const { name, geometry } = feature
+    const { monthlyData } = useClimateData(geometry)
     const title = (name ? `${name} - ` : '') + i18n.t('Weather & Climate')
-
-    useEffect(() => {
-        getTimeSeries(config, geometry).then(setData)
-    }, [geometry])
 
     return (
         <Modal
@@ -75,9 +52,13 @@ const ClimateModal = () => {
                 </Tabs>
                 <div className={styles.tabContent}>
                     {tab === FORECAST && <Forecast geometry={geometry} />}
-                    {tab === TEMPERATURE && <Temperature data={data} />}
-                    {tab === PRECIPITATION && <Precipitation data={data} />}
-                    {tab === CLIMATE && <ClimateChange data={data} />}
+                    {tab === TEMPERATURE && (
+                        <TemperatureTab data={monthlyData} />
+                    )}
+                    {tab === PRECIPITATION && (
+                        <Precipitation data={monthlyData} />
+                    )}
+                    {tab === CLIMATE && <ClimateChange data={monthlyData} />}
                     {tab === AIR && <AirQuality geometry={geometry} />}
                 </div>
             </ModalContent>
