+++ src/components/climate/styles/Modal.module.css
@@ -1,4 +1,5 @@
 .tabContent {
+    position: relative;
     padding-top: 30px;
     min-height: 635px;
 }
