--- src/components/climate/styles/Shared.module.css
@@ -1,13 +1,0 @@
-.chart {
-    position: relative;
-}
-
-.chartButtons {
-    position: absolute;
-    top: 10px;
-    left: 10px;
-}
-
-.chartButtons button {
-    margin-right: 5px;
-}
